feat(public-site): share gallery preview UI
Add a reusable public gallery preview dialog for project screenshots and daydream images.

Update project cards with centered square covers, image-count badges, aligned hover overlays, and shared project link rendering.

Document the public gallery and project card behavior changes.
