+++ src/app/(web)/projects/[projectTypeId]/_components/PreviewProjectImageDialog.tsx
@@ -1,30 +1,9 @@
 "use client";
 
-import classNames from "classnames";
-import { getSafeImageDimensions } from "@/components/Image/getSafeImageDimensions";
-import ImagePreviewContainer from "@/components/Image/ImagePreviewContainer";
-import SupabaseImage from "@/components/Image/SupabaseImage";
-import { useEffect, useRef } from "react";
-import { MdClose } from "react-icons/md";
-import { useGalleryContext } from "@/context/GalleryContext";
-import {
-  DialogProps,
-  CardBody,
-  CardRoot,
-  Dialog,
-  Window,
-  BaseSkeletonLoader,
-  ExternalLink,
-} from "@/components";
-import { useLoadable } from "@/hooks";
-import { useHotkeys } from "react-hotkeys-hook";
-import {
-  PiCode,
-  PiFigmaLogoLight,
-  PiGlobeHemisphereWestThin,
-} from "react-icons/pi";
+import { DialogProps, FormGroup, Label } from "@/components";
+import { GalleryPreviewDialog } from "@/app/(web)/_components/GalleryPreviewDialog";
 import type { ProjectCardItem } from "@/features/projects/types";
-import { Thumbnail } from "@/components/Image/Thumbnail";
+import { ProjectLinks } from "@/app/(web)/projects/[projectTypeId]/_components/ProjectLinks";
 
 interface PreviewProjectImageDialogProps
   extends Required<Pick<DialogProps, "isOpen" | "onClose">>,
@@ -41,147 +20,35 @@ const PreviewProjectImageDialog = ({
   design_link,
   repository_link,
 }: PreviewProjectImageDialogProps) => {
-  const { selectedItem, list, selectedIndex, setSelectedIndex } =
-    useGalleryContext();
-  const { isLoading, startLoading, endLoading } = useLoadable();
-  const safeImageDimensions = getSafeImageDimensions(selectedItem ?? undefined);
-  const scale = safeImageDimensions.width / 1920;
-  const imageContainerRef = useRef<HTMLDivElement>(null);
-
-  useHotkeys(["esc", "escape"], onClose);
-
-  const onLoadingCompleteHandler = () => {
-    imageContainerRef.current?.scrollTo({
-      top: 0,
-    });
-
-    imageContainerRef.current?.focus();
-
-    endLoading();
-  };
-
-  const onClickThumbnailHandler = (index: number) => {
-    setSelectedIndex(index);
-    startLoading();
-  };
-
-  useEffect(() => {
-    startLoading();
-  }, [selectedIndex, startLoading]);
-
-  if (!selectedItem) return null;
-
   return (
-    <Dialog isOpen={isOpen}>
-      <CardRoot className="h-full md:rounded-md">
-        <CardBody className="relative h-full max-w-[1920px] mx-auto">
-          <div className="absolute right-2 top-2 z-10">
-            <button
-              type="button"
-              aria-label="Close preview"
-              className={classNames(
-                "dark:text-neutral-200 text-neutral-800 outline-none p-2 cursor-pointer"
-              )}
-              onClick={onClose}
-            >
-              <MdClose size={28} />
-            </button>
+    <GalleryPreviewDialog
+      isOpen={isOpen}
+      onClose={onClose}
+      thumbnailKeyPrefix="project-thumbnail"
+      sidebar={
+        <div className="flex flex-col w-full lg:w-80 shrink-0 gap-4">
+          <div className="grow text-right lg:grow-0 lg:text-left">
+            <h3 className="hidden lg:block text-lg font-bold capitalize text-neutral-600 dark:text-white mb-2">
+              Details:
+            </h3>
+
+            <FormGroup>
+              <Label className="!mb-0">Project</Label>
+              <p className="text-neutral-700 dark:text-white text-sm lg:text-lg font-bold">
+                {title}
+              </p>
+            </FormGroup>
           </div>
 
-          <ImagePreviewContainer isNavigationDisabled={isLoading}>
-            <div className="flex flex-col md:flex-row justify-between gap-4 md:gap-12 w-full pr-12">
-              <div className="overflow-x-auto ">
-                <div className="flex flex-row shrink-0 pb-2 lg:pb-0">
-                  {list?.map((item, index) => (
-                    <Thumbnail
-                      key={`thumbnail-${index}-${item.name}`}
-                      className="w-16 md:w-20 aspect-square"
-                      image={item}
-                      isActive={index === selectedIndex}
-                      onSelect={() => onClickThumbnailHandler(index)}
-                    />
-                  ))}
-                </div>
-              </div>
-
-              <div className="dark:text-neutral-200 flex flex-col md:justify-start items-start md:items-end flex-1 gap-2 mb-3 md:mb-0">
-                <p className="text-base sm:text-lg block whitespace-nowrap font-bold dark:text-neutral-200 text-neutral-800 font-mono">
-                  {title}
-                </p>
-
-                <div className="flex flex-row gap-4 sm:justify-end">
-                  {repository_link && (
-                    <ExternalLink
-                      href={`https://github.com/${repository_link}`}
-                      label="Code"
-                      className="transform hover:scale-110 cursor-pointer transition-colors text-neutral-800 dark:text-neutral-200"
-                      data-tooltip-id="guest-tooltip"
-                      data-tooltip-content="Code"
-                      data-tooltip-place="bottom"
-                    >
-                      <PiCode size={22} />
-                    </ExternalLink>
-                  )}
-
-                  {website_link && (
-                    <ExternalLink
-                      href={`https://${website_link}`}
-                      label="Link"
-                      className="transform hover:scale-110 cursor-pointer transition-colors text-neutral-800 dark:text-neutral-200"
-                      data-tooltip-id="guest-tooltip"
-                      data-tooltip-content="Website"
-                      data-tooltip-place="bottom"
-                    >
-                      <PiGlobeHemisphereWestThin size={22} />
-                    </ExternalLink>
-                  )}
-
-                  {design_link && (
-                    <ExternalLink
-                      href={`https://www.figma.com/file/${design_link}`}
-                      label="Figma"
-                      className="transform hover:scale-110 cursor-pointer transition-colors text-neutral-800 dark:text-neutral-200"
-                      data-tooltip-id="guest-tooltip"
-                      data-tooltip-content="Design"
-                      data-tooltip-place="bottom"
-                    >
-                      <PiFigmaLogoLight size={22} />
-                    </ExternalLink>
-                  )}
-                </div>
-              </div>
-            </div>
-
-            <div className="relative flex-1">
-              <Window title={selectedItem.name}>
-                <div className="relative h-[calc(100vh-258px)] md:h-[calc(100vh-206px)]">
-                  {isLoading && (
-                    <div className="bg-white absolute inset-0 z-10">
-                      <BaseSkeletonLoader className="w-full h-full" />
-                    </div>
-                  )}
-
-                  <div
-                    className="h-full overflow-y-auto relative scrollbar-w-2 scrollbar scrollbar-thumb-zinc-400 scrollbar-track-gray-200 max-h-[calc(100vh-258px)] md:max-h-[calc(100vh-206px)]"
-                    ref={imageContainerRef}
-                    tabIndex={-1}
-                  >
-                    <SupabaseImage
-                      src={selectedItem.storage_file_path}
-                      alt={selectedItem.name}
-                      width={safeImageDimensions.width / scale}
-                      height={safeImageDimensions.height / scale}
-                      onLoad={onLoadingCompleteHandler}
-                      sizes="100vw"
-                    />
-                  </div>
-                </div>
-              </Window>
-            </div>
-          </ImagePreviewContainer>
-        </CardBody>
-      </CardRoot>
-    </Dialog>
+          <ProjectLinks
+            className="flex flex-row gap-4 justify-end lg:justify-start"
+            design_link={design_link}
+            repository_link={repository_link}
+            website_link={website_link}
+          />
+        </div>
+      }
+    />
   );
 };
